Refactor into a single package.
