--- src/Rovu/IStrategy.java
@@ -1,9 +1,0 @@
-package Rovu;
-
-import javafx.geometry.Point3D;
-
-import java.util.Set;
-
-public interface IStrategy {
-    public Set<Point3D> ComputeStrategy(Set<Point3D> mission);
-}
