+++ src/mdsd/IRobot.java
@@ -1,0 +1,12 @@
+package mdsd;
+
+import project.Point;
+import java.util.Set;
+
+public interface IRobot {
+    public void setMission(Set<Point> positionSet);
+
+    public Point getPosition();
+
+    public void setPosition(Point position);
+}
