+++ src/mdsd/IStrategy.java
@@ -1,0 +1,9 @@
+package mdsd;
+
+import project.Point;
+
+import java.util.Set;
+
+public interface IStrategy {
+    public Set<Point> ComputeStrategy(Set<Point> mission);
+}
